pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "disinfo-relation-checker"
description = "Check if a relation is disinformation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
]
license = {text = "MIT"}
dynamic = ["version"]

[project.scripts]
disinfo-relation-checker = "disinfo_relation_checker.cli:main"


[tool.hatch.build.targets.wheel]
packages = ["src/disinfo_relation_checker"]
py.typed = true

[tool.hatch.version]
source = "regex"
pattern = '^__version__ = "(?P<version>.+)"$'
path = "src/disinfo_relation_checker/__init__.py"

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "nox>=2025.5.1",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.ruff]
target-version = "py311"
line-length = 120
lint.select = ["ALL"]
lint.ignore = ["D203", "D213", "COM812", "PLR2004", "ANN401", "ARG002", "BLE001", "TRY301", "PLC0415", "T201", "S311", "SIM102", "PLW1641", "PLR0913", "B007", "FBT001", "FBT002"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = []
"tests/*" = [
    "S101",
    "S108",
    "S603",
    "B017",
    "PT011",
    "PERF401",
    "SLF001",
]
"src/disinfo_relation_checker/prompt_optimizer.py" = [
    "S101",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = false
strict_equality = true

[tool.pytest.ini_options]
addopts = ["-sv", "--doctest-modules"]
testpaths = [
    "tests",
    "src/disinfo_relation_checker",
]
pythonpath = [
    "src",
]
filterwarnings = [
    "error",
]
markers = [
    "e2e: marks tests as end-to-end tests"
]
